Add spacing, abstract checkbox listener
