+++ client/src/components/RegisterVolunteer/CodeOfConduct.jsx
@@ -25,64 +25,64 @@ export default function CodeOfConduct({ setCanSubmit }) {
       </Label3>
 
       <CocContainer>
-      <BodyText2>
+        <BodyText2>
             Adapted from the{' '}
             <a href="https://github.com/codeforamerica/codeofconduct">
               CfA Code of Conduct
             </a>.
+        </BodyText2>
 
-            <p>We are an official brigade. The Code for Chicago community expects that Code for Chicago
-            events, and digital forums:</p>
+            <BodyText2>We are an official brigade. The Code for Chicago community expects that Code for Chicago
+            events, and digital forums:</BodyText2>
 
 
-          <p>Are a safe and respectful environment for all participants.</p>
-          <p>
+          <BodyText2>Are a safe and respectful environment for all participants.</BodyText2>
+          <BodyText2>
             Are a place where people are free to fully express their identities.
-          </p>
-          <p>
+          </BodyText2>
+          <BodyText2>
             Presume the value of others. Everyone&apos;s ideas, skills, and contributions have
             value.
-          </p>
-          <p>
+          </BodyText2>
+          <BodyText2>
             Don&apos;t assume everyone has the same context, and encourage questions.
-          </p>
-          <p>
+          </BodyText2>
+          <BodyText2>
             Find a way for people to be productive with their skills (technical and not) and energy.
             Use language such as “yes/and”, not “no/but.”
-          </p>
-          <p>
+          </BodyText2>
+          <BodyText2>
             Encourage members and participants to listen as much as they speak.
-          </p>
-          <p>
+          </BodyText2>
+          <BodyText2>
             Strive to build tools that are open and free technology for public use. Activities that
             aim to foster public use, not private gain, are prioritized.
-          </p>
-          <p>
+          </BodyText2>
+          <BodyText2>
             Prioritize access for and input from those who are traditionally excluded from the civic
             process.
-          </p>
-          <p>
+          </BodyText2>
+          <BodyText2>
             Work to ensure that the community is well-represented in the planning, design, and
             implementation of civic tech. This includes encouraging participation from women,
             minorities, and traditionally marginalized groups.{' '}
-          </p>
-          <p>
+          </BodyText2>
+          <BodyText2>
             Actively involve community groups and those with subject matter expertise in the
             decision-making process.
-          </p>
-          <p>
+          </BodyText2>
+          <BodyText2>
             Ensure that the relationships and conversations between community members, the local
             government staff and community partners remain respectful, participatory, and
             productive.
-          </p>
-          <p>
+          </BodyText2>
+          <BodyText2>
             Provide an environment where people are free from discrimination or harassment.
-          </p>
+          </BodyText2>
 
-          <p>
+          <BodyText2>
             Code for Chicago reserves the right to ask anyone in violation of these policies not to
             participate in Code for Chicago network activities, events, and digital forums.
-          </p>
         </BodyText2>
         </CocContainer>
         <Label3>
